Split to core.rs
