+++ src/consts.rs
@@ -1,0 +1,1 @@
+pub const EXIT_CODE_ERROR: i32 = 2;
